+++ packages/dev-cypress/js/helpers/responsive.js
@@ -1,5 +1,7 @@
 export function setDeviceType(deviceType) {
-	cy.getByAriaLabel('Breakpoints').within(() => {
-		cy.getByAriaLabel(deviceType).click({ force: true });
-	});
+	cy.getByAriaLabel('Breakpoints')
+		.first()
+		.within(() => {
+			cy.getByAriaLabel(deviceType).click({ force: true });
+		});
 }